server: Run e2e tests on Docker Compose
Change-type: patch
