+++ .circleci/config.yml
@@ -106,7 +106,7 @@ jobs:
 
       - run:
           name: Integration Tests
-          command: make test-integration COVERAGE=1
+          command: make test-integration-core test-integration-queue test-integration-sync test-integration-worker COVERAGE=1 && make test-integration-server COVERAGE=1 INTEGRATION_OUTREACH_APP_ID=$OUTREACH_APP_ID INTEGRATION_OUTREACH_APP_SECRET=$OUTREACH_APP_SECRET INTEGRATION_OUTREACH_SIGNATURE_KEY=$OUTREACH_SIGNATURE_KEY OAUTH_REDIRECT_BASE_URL=https://jel.ly.fish
       - run:
           name: Integration Tests (with debug logs)
           command: make test-integration COVERAGE=1 LOGLEVEL=debug
@@ -279,10 +279,10 @@ jobs:
 
       - run:
           name: Sync Tests
-          command: make test FILES=./test/integration/sync/outreach-mirror.spec.js COVERAGE=1 INTEGRATION_OUTREACH_APP_ID=$OUTREACH_APP_ID INTEGRATION_OUTREACH_APP_SECRET=$OUTREACH_APP_SECRET INTEGRATION_OUTREACH_SIGNATURE_KEY=$OUTREACH_SIGNATURE_KEY
+          command: make test FILES=./test/integration/server/outreach-mirror.spec.js COVERAGE=1 INTEGRATION_OUTREACH_APP_ID=$OUTREACH_APP_ID INTEGRATION_OUTREACH_APP_SECRET=$OUTREACH_APP_SECRET INTEGRATION_OUTREACH_SIGNATURE_KEY=$OUTREACH_SIGNATURE_KEY
       - run:
           name: Sync End To End Tests Without Tokens
-          command: make test FILES=./test/integration/sync/outreach-mirror.spec.js COVERAGE=1 INTEGRATION_OUTREACH_APP_ID= INTEGRATION_OUTREACH_APP_SECRET= INTEGRATION_OUTREACH_SIGNATURE_KEY=
+          command: make test FILES=./test/integration/server/outreach-mirror.spec.js COVERAGE=1 INTEGRATION_OUTREACH_APP_ID= INTEGRATION_OUTREACH_APP_SECRET= INTEGRATION_OUTREACH_SIGNATURE_KEY=
 
       - persist_to_workspace:
           root: .
@@ -388,39 +388,26 @@ jobs:
           docker_layer_caching: true
 
       - run:
-          name: End To End Tests (Docker Compose)
-          command: make compose-up DETACH=1 && make compose-exec-sidecar COMMAND="make" ARGS='test COVERAGE=0 FILES=./test/e2e/sdk/*.spec.js SCRUB=0' && make compose-exec-sidecar COMMAND="make" ARGS="test-e2e-ui COVERAGE=0 SCRUB=0" && make compose-down
-
+          name: Docker Compose Up (with integration tokens)
+          command: make compose-up DETACH=1 INTEGRATION_GITHUB_TOKEN=$GITHUB_TOKEN INTEGRATION_GITHUB_SIGNATURE_KEY=$GITHUB_SIGNATURE_KEY INTEGRATION_OUTREACH_APP_ID=$OUTREACH_APP_ID INTEGRATION_OUTREACH_APP_SECRET=$OUTREACH_APP_SECRET INTEGRATION_OUTREACH_SIGNATURE_KEY=$OUTREACH_SIGNATURE_KEY INTEGRATION_FLOWDOCK_SIGNATURE_KEY=$FLOWDOCK_SIGNATURE_KEY OAUTH_REDIRECT_BASE_URL=https://jel.ly.fish
       - run:
-          name: Start PostgreSQL
-          command: service postgresql start && su - postgres -c "psql -U postgres -d postgres -c \"alter user postgres with password 'postgres';\""
+          name: End To End Tests (Server)
+          command: make compose-exec-sidecar COMMAND="make" ARGS="test-e2e-server COVERAGE=0 SCRUB=0 INTEGRATION_GITHUB_TOKEN=$GITHUB_TOKEN INTEGRATION_GITHUB_SIGNATURE_KEY=$GITHUB_SIGNATURE_KEY INTEGRATION_OUTREACH_APP_ID=$OUTREACH_APP_ID INTEGRATION_OUTREACH_APP_SECRET=$OUTREACH_APP_SECRET INTEGRATION_OUTREACH_SIGNATURE_KEY=$OUTREACH_SIGNATURE_KEY INTEGRATION_FLOWDOCK_SIGNATURE_KEY=$FLOWDOCK_SIGNATURE_KEY OAUTH_REDIRECT_BASE_URL=https://jel.ly.fish"
       - run:
-          name: Start Redis
-          command: service redis-server start
-
-      # Setup artifact store for UI integration test screenshots
-      - run: mkdir -p /tmp/test-results/screenshots
-
+          name: Docker Compose Down
+          command: make compose-down
       - run:
-          name: End To End Tests
-          command: make test-e2e-server COVERAGE=1 INTEGRATION_GITHUB_TOKEN=$GITHUB_TOKEN INTEGRATION_GITHUB_SIGNATURE_KEY=$GITHUB_SIGNATURE_KEY INTEGRATION_OUTREACH_APP_ID=$OUTREACH_APP_ID INTEGRATION_OUTREACH_APP_SECRET=$OUTREACH_APP_SECRET INTEGRATION_OUTREACH_SIGNATURE_KEY=$OUTREACH_SIGNATURE_KEY INTEGRATION_FLOWDOCK_SIGNATURE_KEY=$FLOWDOCK_SIGNATURE_KEY
-
+          name: Docker Compose Up
+          command: make compose-up DETACH=1
       - run:
-          name: Postgres Dump
-          command: ./scripts/postgres-dump.sh dump.gz
-
-      # Store artifacts for debugging purposes
-      - store_test_results:
-          path: /tmp/test-results
-      - store_artifacts:
-          path: /tmp/test-results
-          destination: test-results
-
-      - persist_to_workspace:
-          root: .
-          paths:
-            - .nyc_output/*.json
-            - dump.gz
+          name: End To End Tests (SDK)
+          command: make compose-exec-sidecar COMMAND="make" ARGS="test-e2e-sdk COVERAGE=0 SCRUB=0"
+      - run:
+          name: End To End Tests (UI)
+          command: make compose-exec-sidecar COMMAND="make" ARGS="test-e2e-ui COVERAGE=0 SCRUB=0"
+      - run:
+          name: Docker Compose Down
+          command: make compose-down
 
   results:
     <<: *job_defaults
@@ -431,9 +418,9 @@ jobs:
       - attach_workspace:
           at: .
 
-      - run:
-          name: Postgres Dump
-          command: mkdir -p /tmp/test-results && cp dump.gz /tmp/test-results
+      # - run:
+          # name: Postgres Dump
+          # command: mkdir -p /tmp/test-results && cp dump.gz /tmp/test-results
 
       - run:
           name: Webpack Bundle Report For UI
