+++ lib/action-library/index.js
@@ -1045,14 +1045,14 @@ module.exports = {
 				}
 
 				try {
-					return module.exports['action-create-card'].handler(
+					return await module.exports['action-create-card'].handler(
 						session,
 						context,
 						card,
 						createOptions)
 				} catch (error) {
 					// Notice action-create-card throws an error if the card
-					// you wan to create already exists. Because we check if
+					// you want to create already exists. Because we check if
 					// the tag exists to decide whether to update or insert in
 					// a non atomic way, two calls can concurrently think the
 					// tag doesn't exist, and therefore one will fail.
@@ -1061,10 +1061,16 @@ module.exports = {
 					// can check if our insert failed, and if so retry using
 					// an update instead.
 					if (error.name === 'JellyfishElementAlreadyExists') {
+						// Get the card again
+						const input = await context.getCardBySlug(
+							session, slug, {
+								type: 'tag'
+							})
+
 						return module.exports['action-increment'].handler(
 							session,
 							context,
-							tagCard,
+							input,
 							incrementOptions)
 					}
