+++ test/e2e/server/hooks.spec.js
@@ -6,9 +6,26 @@
 
 const ava = require('ava')
 const helpers = require('../sdk/helpers')
+const environment = require('../../../lib/environment')
 
-ava.before(helpers.sdk.before)
-ava.after(helpers.sdk.after)
+ava.before(async (test) => {
+	await helpers.before(test)
+
+	const session = await test.context.sdk.auth.login({
+		username: environment.test.user.username,
+		password: environment.test.user.password
+	})
+
+	test.context.token = session.id
+})
+
+ava.after(helpers.after)
+
+ava.beforeEach(async (test) => {
+	await helpers.beforeEach(test, test.context.token)
+})
+
+ava.afterEach(helpers.afterEach)
 
 ava.serial('should post a dummy "none" event', async (test) => {
 	const result = await test.context.http('POST', '/api/v2/hooks/none', {
