+++ test/e2e/server/messages.spec.js
@@ -9,7 +9,7 @@ const helpers = require('../sdk/helpers')
 const environment = require('../../../lib/environment')
 
 ava.before(async (test) => {
-	await helpers.sdk.before(test)
+	await helpers.before(test)
 
 	const session = await test.context.sdk.auth.login({
 		username: environment.test.user.username,
@@ -19,13 +19,13 @@ ava.before(async (test) => {
 	test.context.token = session.id
 })
 
-ava.after(helpers.sdk.after)
+ava.after(helpers.after)
 
 ava.beforeEach(async (test) => {
-	await helpers.sdk.beforeEach(test, test.context.token)
+	await helpers.beforeEach(test, test.context.token)
 })
 
-ava.afterEach(helpers.sdk.afterEach)
+ava.afterEach(helpers.afterEach)
 
 ava.serial('should sanely handle line breaks before tags in messages/whispers', async (test) => {
 	const thread = await test.context.sdk.card.create({
