+++ test/e2e/server/oauth.spec.js
@@ -6,9 +6,26 @@
 
 const ava = require('ava')
 const helpers = require('../sdk/helpers')
+const environment = require('../../../lib/environment')
 
-ava.before(helpers.sdk.before)
-ava.after(helpers.sdk.after)
+ava.before(async (test) => {
+	await helpers.before(test)
+
+	const session = await test.context.sdk.auth.login({
+		username: environment.test.user.username,
+		password: environment.test.user.password
+	})
+
+	test.context.token = session.id
+})
+
+ava.after(helpers.after)
+
+ava.beforeEach(async (test) => {
+	await helpers.beforeEach(test, test.context.token)
+})
+
+ava.afterEach(helpers.afterEach)
 
 ava.serial('/api/v2/oauth should return 400 given an unknown oauth integration', async (test) => {
 	const result = await test.context.http(
