+++ test/e2e/server/users.spec.js
@@ -12,7 +12,7 @@ const helpers = require('../sdk/helpers')
 const environment = require('../../../lib/environment')
 
 ava.before(async (test) => {
-	await helpers.sdk.before(test)
+	await helpers.before(test)
 
 	const session = await test.context.sdk.auth.login({
 		username: environment.test.user.username,
@@ -22,13 +22,13 @@ ava.before(async (test) => {
 	test.context.token = session.id
 })
 
-ava.after(helpers.sdk.after)
+ava.after(helpers.after)
 
 ava.beforeEach(async (test) => {
-	await helpers.sdk.beforeEach(test, test.context.token)
+	await helpers.beforeEach(test, test.context.token)
 })
 
-ava.afterEach(helpers.sdk.afterEach)
+ava.afterEach(helpers.afterEach)
 
 const createUserDetails = () => {
 	return {
